+++ frontend/components/event/event_index_container.jsx
