+++ frontend/components/event/event_index_item.jsx
