+++ frontend/components/event/event_show_container.jsx
