+++ frontend/components/group/group_index_container.jsx
