+++ frontend/components/main/footer.jsx
