+++ frontend/components/main/homepage.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import GroupSearchContainer from './group_search/group_search_container';
-import EventIndexContainer from './event_index/event_index_container';
-import CategoryIndexContainer from './category_index/category_index_container';
-import * as SplashUtil from '../util/splash_util';
+import GroupSearchContainer from '../group/group_search_container';
+import EventIndexContainer from '../event/event_index_container';
+import CategoryIndexContainer from '../category/category_index_container';
+import * as SplashUtil from '../../util/splash_util';
 
 
 class Homepage extends React.Component {
