+++ frontend/components/xyz.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
-import NavBarContainer from './navbar/navbar_container';
+import NavBarContainer from './main/navbar_container';
 // import SessionFormContainer from './session_form/session_form_container';
 // import GroupIndexContainer from './group_index/group_index_container';
-import GroupShowContainer from './group_show/group_show_container';
-import EventShowContainer from './event_show/event_show_container';
+import GroupShowContainer from './group/group_show_container';
+import EventShowContainer from './event/event_show_container';
 // import EventIndexContainer from './event_index/event_index_container';
-import CategoryShowContainer from './category_show/category_show_container';
-import Homepage from './homepage';
-import Discover from './discover';
-import FooterContainer from './footer/footer_container';
+import CategoryShowContainer from './category/category_show_container';
+import Homepage from './main/homepage';
+import Discover from './main/discover';
+import FooterContainer from './main/footer_container';
 
 const App = () => (
   <div>
